backend/requirements.txt
appdirs==1.4.4
astroid==2.2.5
attrs==19.3.0
cached-property==1.5.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.1
defusedxml==0.6.0
Django==2.2.3
django-cors-headers==3.2.0
djangorestframework==3.10.1
idna==2.9
isodate==0.6.0
isort==4.3.21
lazy-object-proxy==1.4.1
lxml==4.5.1
mccabe==0.6.1
pylint==2.3.1
pytz==2019.1
requests==2.24.0
requests-toolbelt==0.9.1
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.9
virtualenv==16.7.7
wrapt==1.11.2
zeep==3.4.0
gunicorn
mysqlclient
